✨ update Props 정렬
